refactor: fold assistant runtime into agent runtime
