+++ src-tauri/crates/tauri-plugin-agent-runtime/src/audit.rs
