+++ src-tauri/crates/tauri-plugin-agent-runtime/src/catalog.rs
