+++ src-tauri/crates/tauri-plugin-agent-runtime/src/commands.rs
