+++ src-tauri/crates/tauri-plugin-agent-runtime/src/error.rs
