+++ src-tauri/crates/tauri-plugin-agent-runtime/src/events.rs
