+++ src-tauri/crates/tauri-plugin-agent-runtime/src/lib.rs
@@ -1,11 +1,25 @@
-//! `tauri-plugin-agent-runtime` 是框架对外统一的 plugin 入口。
+//! `tauri-plugin-agent-runtime` 是框架对外统一的 Tauri plugin 入口。
 //!
-//! 新宿主应用应直接依赖本 crate，并使用 `agent-runtime` namespace。
-//! `tauri-plugin-assistant-runtime` 只作为仓库内部实现承载 crate 存在，
-//! 不应被新的宿主应用直接当作公共入口使用。
+//! 运行时实现、thread/turn/stream 主链、native tools、proposal 与
+//! catalog/wiring 全部收敛在本 crate 中，对外统一暴露 `agent-runtime`
+//! namespace。
 
-use tauri::plugin::TauriPlugin;
-use tauri::Runtime;
+mod audit;
+mod catalog;
+mod commands;
+mod config;
+mod db;
+mod error;
+mod events;
+mod models;
+mod native_tools;
+mod proposal;
+mod runtime;
+mod session;
+
+use serde_json::json;
+use tauri::plugin::{Builder as PluginBuilder, TauriPlugin};
+use tauri::{AppHandle, Manager, Runtime};
 
 pub const PLUGIN_NAME: &str = "agent-runtime";
 
@@ -16,30 +30,145 @@ pub use agent_core::{
     RuntimeConfig, SkillBinding, SkillCatalogRegistration, ToolBinding, ToolProvider, UiMetadata,
     Visibility,
 };
-pub use tauri_plugin_assistant_runtime::{
-    interrupt_turn, resolve_runtime_request, start_headless_turn, AssistantRuntimeState,
-    AttachmentRef, ConfigSelection, Error, Result, StartedTurn, StreamDispatch,
-};
-pub use tauri_plugin_assistant_runtime::{
-    AssistantStatusEventEnvelope, AssistantTurnStreamEnvelope, CancelTurnAck,
-    ConfigFileCandidate, ConversationBlock, CreateThreadResponse, ListConfigFilesResponse,
-    ListThreadsResponse, NativeToolAuditEntry, PendingRuntimeRequest, ResolvedConfig,
-    ResumeTurnStreamAck, RuntimeCatalog, RuntimeCatalogIntegration, RuntimeCatalogSkill,
-    RuntimeCatalogTool, StartTurnAck, SubmitRuntimeRequestAck, ThreadSnapshot, ThreadSummary,
-    TurnSummarySnapshot,
+pub use error::{Error, Result};
+pub use models::*;
+pub use session::{AssistantRuntimeState, StartedTurn, StreamDispatch};
+pub type AgentRuntimeState = AssistantRuntimeState;
+pub type AgentStatusEventEnvelope = AssistantStatusEventEnvelope;
+pub type AgentTurnStreamEnvelope = AssistantTurnStreamEnvelope;
+
+#[derive(Clone, Copy)]
+pub(crate) struct RuntimePluginMetadata {
+    pub(crate) status_event: &'static str,
+    pub(crate) threads_changed_event: &'static str,
+    pub(crate) debug_event: &'static str,
+    pub(crate) audit_directory: &'static str,
+}
+
+pub(crate) const RUNTIME_PLUGIN_METADATA: RuntimePluginMetadata = RuntimePluginMetadata {
+    status_event: "agent-runtime://status",
+    threads_changed_event: "agent-runtime://threads-changed",
+    debug_event: "agent-runtime://debug",
+    audit_directory: PLUGIN_NAME,
 };
 
+pub(crate) fn runtime_plugin_metadata<R: Runtime>(_app: &AppHandle<R>) -> RuntimePluginMetadata {
+    RUNTIME_PLUGIN_METADATA
+}
+
+pub async fn start_headless_turn<R: Runtime>(
+    app: AppHandle<R>,
+    state: AssistantRuntimeState,
+    thread_id: String,
+    text: String,
+    mode: String,
+    attachments: Option<Vec<AttachmentRef>>,
+    model_override: Option<String>,
+    config_context: Option<ConfigSelection>,
+) -> Result<StartTurnAck> {
+    let attachments = attachments.unwrap_or_default();
+    let started = state.start_background_turn(&thread_id, &text, config_context.clone())?;
+    let turn_id = started.turn_id.clone();
+    let accepted_at = started.accepted_at.clone();
+    let resolved_config_context = state.thread_config_selection(&thread_id)?;
+
+    match runtime::start_runtime_turn(
+        app.clone(),
+        state.clone(),
+        thread_id.clone(),
+        turn_id.clone(),
+        text,
+        attachments,
+        mode,
+        model_override,
+        resolved_config_context,
+    )
+    .await
+    {
+        Ok(()) => {
+            events::emit_threads_changed(&app, "updated", Some(&thread_id));
+            Ok(StartTurnAck {
+                thread_id,
+                turn_id,
+                accepted_at,
+            })
+        }
+        Err(err) => {
+            let StreamDispatch { item, subscribers } = state.push_stream_event(
+                &thread_id,
+                &turn_id,
+                "plugin",
+                "turn_failed",
+                json!({
+                    "code": "runtime_start_failed",
+                    "message": err.to_string(),
+                }),
+            )?;
+            for subscriber in subscribers {
+                let _ = subscriber.send(item.clone());
+            }
+            events::emit_threads_changed(&app, "updated", Some(&thread_id));
+            Err(err)
+        }
+    }
+}
+
+pub async fn resolve_runtime_request<R: Runtime>(
+    app: AppHandle<R>,
+    state: AssistantRuntimeState,
+    thread_id: String,
+    turn_id: String,
+    request_id: String,
+    response: serde_json::Value,
+) -> Result<String> {
+    runtime::submit_runtime_request_response(app, state, thread_id, turn_id, request_id, response)
+        .await
+}
+
+pub async fn interrupt_turn<R: Runtime>(
+    app: AppHandle<R>,
+    state: AssistantRuntimeState,
+    thread_id: String,
+    turn_id: String,
+) -> Result<bool> {
+    runtime::interrupt_runtime_turn(app, state, thread_id, turn_id).await
+}
+
 pub fn init<R: Runtime>() -> TauriPlugin<R> {
-    tauri_plugin_assistant_runtime::init_with_name(PLUGIN_NAME)
+    PluginBuilder::new(PLUGIN_NAME)
+        .invoke_handler(tauri::generate_handler![
+            commands::list_threads,
+            commands::get_thread_snapshot,
+            commands::create_thread,
+            commands::start_turn,
+            commands::resume_turn_stream,
+            commands::cancel_turn,
+            commands::submit_runtime_request,
+            commands::list_config_files,
+            commands::resolve_config_profile,
+            commands::get_runtime_catalog,
+        ])
+        .setup(|app, _api| {
+            app.manage(AssistantRuntimeState::default());
+            Ok(())
+        })
+        .build()
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use super::{Builder, PLUGIN_NAME, RUNTIME_PLUGIN_METADATA};
 
     #[test]
-    fn generic_plugin_uses_agent_runtime_namespace() {
+    fn agent_runtime_metadata_uses_agent_namespace() {
         assert_eq!(PLUGIN_NAME, "agent-runtime");
+        assert_eq!(RUNTIME_PLUGIN_METADATA.status_event, "agent-runtime://status");
+        assert_eq!(
+            RUNTIME_PLUGIN_METADATA.threads_changed_event,
+            "agent-runtime://threads-changed"
+        );
+        assert_eq!(RUNTIME_PLUGIN_METADATA.debug_event, "agent-runtime://debug");
+        assert_eq!(RUNTIME_PLUGIN_METADATA.audit_directory, "agent-runtime");
         let _ = Builder::new();
     }
 }
