+++ src-tauri/crates/tauri-plugin-agent-runtime/src/native_tools.rs
