+++ src-tauri/crates/tauri-plugin-agent-runtime/src/session.rs
