--- src-tauri/crates/tauri-plugin-assistant-runtime/permissions/autogenerated/commands/cancel_turn.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-cancel-turn"
-description = "Enables the cancel_turn command without any pre-configured scope."
-commands.allow = ["cancel_turn"]
-
-[[permission]]
-identifier = "deny-cancel-turn"
-description = "Denies the cancel_turn command without any pre-configured scope."
-commands.deny = ["cancel_turn"]
