--- src-tauri/crates/tauri-plugin-assistant-runtime/permissions/autogenerated/commands/create_thread.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-create-thread"
-description = "Enables the create_thread command without any pre-configured scope."
-commands.allow = ["create_thread"]
-
-[[permission]]
-identifier = "deny-create-thread"
-description = "Denies the create_thread command without any pre-configured scope."
-commands.deny = ["create_thread"]
