--- src-tauri/crates/tauri-plugin-assistant-runtime/permissions/autogenerated/commands/get_thread_snapshot.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-get-thread-snapshot"
-description = "Enables the get_thread_snapshot command without any pre-configured scope."
-commands.allow = ["get_thread_snapshot"]
-
-[[permission]]
-identifier = "deny-get-thread-snapshot"
-description = "Denies the get_thread_snapshot command without any pre-configured scope."
-commands.deny = ["get_thread_snapshot"]
