--- src-tauri/crates/tauri-plugin-assistant-runtime/permissions/autogenerated/commands/list_config_files.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-list-config-files"
-description = "Enables the list_config_files command without any pre-configured scope."
-commands.allow = ["list_config_files"]
-
-[[permission]]
-identifier = "deny-list-config-files"
-description = "Denies the list_config_files command without any pre-configured scope."
-commands.deny = ["list_config_files"]
