--- src-tauri/crates/tauri-plugin-assistant-runtime/permissions/autogenerated/commands/list_threads.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-list-threads"
-description = "Enables the list_threads command without any pre-configured scope."
-commands.allow = ["list_threads"]
-
-[[permission]]
-identifier = "deny-list-threads"
-description = "Denies the list_threads command without any pre-configured scope."
-commands.deny = ["list_threads"]
