--- src-tauri/crates/tauri-plugin-assistant-runtime/permissions/autogenerated/commands/resolve_config_profile.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-resolve-config-profile"
-description = "Enables the resolve_config_profile command without any pre-configured scope."
-commands.allow = ["resolve_config_profile"]
-
-[[permission]]
-identifier = "deny-resolve-config-profile"
-description = "Denies the resolve_config_profile command without any pre-configured scope."
-commands.deny = ["resolve_config_profile"]
