--- src-tauri/crates/tauri-plugin-assistant-runtime/permissions/autogenerated/commands/resume_turn_stream.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-resume-turn-stream"
-description = "Enables the resume_turn_stream command without any pre-configured scope."
-commands.allow = ["resume_turn_stream"]
-
-[[permission]]
-identifier = "deny-resume-turn-stream"
-description = "Denies the resume_turn_stream command without any pre-configured scope."
-commands.deny = ["resume_turn_stream"]
