--- src-tauri/crates/tauri-plugin-assistant-runtime/permissions/autogenerated/commands/start_turn.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-start-turn"
-description = "Enables the start_turn command without any pre-configured scope."
-commands.allow = ["start_turn"]
-
-[[permission]]
-identifier = "deny-start-turn"
-description = "Denies the start_turn command without any pre-configured scope."
-commands.deny = ["start_turn"]
