--- src-tauri/crates/tauri-plugin-assistant-runtime/permissions/autogenerated/commands/submit_runtime_request.toml
@@ -1,13 +1,0 @@
-# Automatically generated - DO NOT EDIT!
-
-"$schema" = "../../schemas/schema.json"
-
-[[permission]]
-identifier = "allow-submit-runtime-request"
-description = "Enables the submit_runtime_request command without any pre-configured scope."
-commands.allow = ["submit_runtime_request"]
-
-[[permission]]
-identifier = "deny-submit-runtime-request"
-description = "Denies the submit_runtime_request command without any pre-configured scope."
-commands.deny = ["submit_runtime_request"]
