--- src-tauri/crates/tauri-plugin-assistant-runtime/permissions/autogenerated/reference.md
@@ -1,286 +1,0 @@
-## 默认权限
-
-嵌入式 assistant runtime bridge 的默认权限。
-
-#### 默认权限集包含：
-
-- `allow-list-threads`
-- `allow-get-thread-snapshot`
-- `allow-create-thread`
-- `allow-start-turn`
-- `allow-resume-turn-stream`
-- `allow-cancel-turn`
-- `allow-submit-runtime-request`
-- `allow-list-config-files`
-- `allow-resolve-config-profile`
-- `allow-get-runtime-catalog`
-
-## 权限表
-
-<table>
-<tr>
-<th>Identifier</th>
-<th>Description</th>
-</tr>
-
-
-<tr>
-<td>
-
-`assistant-runtime:allow-cancel-turn`
-
-</td>
-<td>
-
-允许在无预设作用域的情况下调用 cancel_turn 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:deny-cancel-turn`
-
-</td>
-<td>
-
-禁止在无预设作用域的情况下调用 cancel_turn 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:allow-create-thread`
-
-</td>
-<td>
-
-允许在无预设作用域的情况下调用 create_thread 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:deny-create-thread`
-
-</td>
-<td>
-
-禁止在无预设作用域的情况下调用 create_thread 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:allow-get-runtime-catalog`
-
-</td>
-<td>
-
-允许在无预设作用域的情况下调用 get_runtime_catalog 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:deny-get-runtime-catalog`
-
-</td>
-<td>
-
-禁止在无预设作用域的情况下调用 get_runtime_catalog 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:allow-get-thread-snapshot`
-
-</td>
-<td>
-
-允许在无预设作用域的情况下调用 get_thread_snapshot 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:deny-get-thread-snapshot`
-
-</td>
-<td>
-
-禁止在无预设作用域的情况下调用 get_thread_snapshot 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:allow-list-config-files`
-
-</td>
-<td>
-
-允许在无预设作用域的情况下调用 list_config_files 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:deny-list-config-files`
-
-</td>
-<td>
-
-禁止在无预设作用域的情况下调用 list_config_files 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:allow-list-threads`
-
-</td>
-<td>
-
-允许在无预设作用域的情况下调用 list_threads 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:deny-list-threads`
-
-</td>
-<td>
-
-禁止在无预设作用域的情况下调用 list_threads 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:allow-resolve-config-profile`
-
-</td>
-<td>
-
-允许在无预设作用域的情况下调用 resolve_config_profile 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:deny-resolve-config-profile`
-
-</td>
-<td>
-
-禁止在无预设作用域的情况下调用 resolve_config_profile 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:allow-resume-turn-stream`
-
-</td>
-<td>
-
-允许在无预设作用域的情况下调用 resume_turn_stream 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:deny-resume-turn-stream`
-
-</td>
-<td>
-
-禁止在无预设作用域的情况下调用 resume_turn_stream 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:allow-start-turn`
-
-</td>
-<td>
-
-允许在无预设作用域的情况下调用 start_turn 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:deny-start-turn`
-
-</td>
-<td>
-
-禁止在无预设作用域的情况下调用 start_turn 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:allow-submit-runtime-request`
-
-</td>
-<td>
-
-允许在无预设作用域的情况下调用 submit_runtime_request 命令。
-
-</td>
-</tr>
-
-<tr>
-<td>
-
-`assistant-runtime:deny-submit-runtime-request`
-
-</td>
-<td>
-
-禁止在无预设作用域的情况下调用 submit_runtime_request 命令。
-
-</td>
-</tr>
-</table>
