--- src-tauri/crates/tauri-plugin-assistant-runtime/permissions/default.toml
@@ -1,18 +1,0 @@
-"$schema" = "schemas/schema.json"
-
-[default]
-description = """
-Default permissions for the embedded assistant runtime bridge.
-"""
-permissions = [
-  "allow-list-threads",
-  "allow-get-thread-snapshot",
-  "allow-create-thread",
-  "allow-start-turn",
-  "allow-resume-turn-stream",
-  "allow-cancel-turn",
-  "allow-submit-runtime-request",
-  "allow-list-config-files",
-  "allow-resolve-config-profile",
-  "allow-get-runtime-catalog",
-]
